Add refactored test files for better coverage analysis
