+++ tests/engine/test_common_engine1.py
@@ -188,331 +188,6 @@ class TestCommonEngine(unittest.TestCase):
         if hasattr(self.engine, "ipc_signal_suffix"):
             self.assertIsNotNone(self.engine.ipc_signal_suffix)
 
-    def test_create_data_processor(self):
-        """Test create_data_processor method exists and is callable"""
-        self.assertTrue(hasattr(self.engine, "create_data_processor"))
-        self.assertTrue(callable(getattr(self.engine, "create_data_processor")))
-
-    def test_init_worker_monitor_signals(self):
-        """Test _init_worker_monitor_signals method exists and is callable"""
-        self.assertTrue(hasattr(self.engine, "_init_worker_monitor_signals"))
-        self.assertTrue(callable(getattr(self.engine, "_init_worker_monitor_signals")))
-
-    def test_start_worker_queue_service(self):
-        """Test start_worker_queue_service method exists and is callable"""
-        self.assertTrue(hasattr(self.engine, "start_worker_queue_service"))
-        self.assertTrue(callable(getattr(self.engine, "start_worker_queue_service")))
-        # Method should not raise exception when called
-        try:
-            # This method should be safe to call multiple times
-            self.engine.start_worker_queue_service()
-        except Exception as e:
-            # In some cases this might fail if already started, which is OK
-            pass
-
-    def test_decode_token(self):
-        """Test _decode_token method exists and is callable"""
-        self.assertTrue(hasattr(self.engine, "_decode_token"))
-        self.assertTrue(callable(getattr(self.engine, "_decode_token")))
-        # Test with sample data (this method should be safe to call)
-        try:
-            token_ids = [1, 2, 3, 4, 5]
-            result = self.engine._decode_token(token_ids, "test_req", False)
-            self.assertIsInstance(result, tuple)
-            self.assertEqual(len(result), 2)
-        except Exception as e:
-            # May fail if data_processor not properly initialized
-            pass
-
-    def test_start_cache_service(self):
-        """Test start_cache_service method exists and is callable"""
-        self.assertTrue(hasattr(self.engine, "start_cache_service"))
-        self.assertTrue(callable(getattr(self.engine, "start_cache_service")))
-        # Method should be safe to call (returns list of processes)
-        try:
-            result = self.engine.start_cache_service([], "test_suffix")
-            self.assertIsInstance(result, list)
-        except Exception as e:
-            # May fail in test environment, which is OK
-            pass
-
-    def test_check_and_free_block_tables(self):
-        """Test check_and_free_block_tables method exists and is callable"""
-        self.assertTrue(hasattr(self.engine, "check_and_free_block_tables"))
-        self.assertTrue(callable(getattr(self.engine, "check_and_free_block_tables")))
-        # Method should be safe to call (just delegates to resource_manager)
-        try:
-            self.engine.check_and_free_block_tables()
-        except Exception as e:
-            # May fail if resource_manager not properly initialized
-            pass
-
-    def test_clear_data(self):
-        """Test clear_data method exists and is callable"""
-        self.assertTrue(hasattr(self.engine, "clear_data"))
-        self.assertTrue(callable(getattr(self.engine, "clear_data")))
-
-    def test_register_to_router(self):
-        """Test _register_to_router method exists and is callable"""
-        self.assertTrue(hasattr(self.engine, "_register_to_router"))
-        self.assertTrue(callable(getattr(self.engine, "_register_to_router")))
-
-    def test_engine_model_config_attributes(self):
-        """Test that engine has expected model configuration attributes"""
-        self.assertTrue(hasattr(self.engine, "cfg"))
-        self.assertIsNotNone(self.engine.cfg)
-
-        if hasattr(self.engine.cfg, "model_config"):
-            model_cfg = self.engine.cfg.model_config
-            self.assertIsNotNone(model_cfg)
-
-            # Test key model config attributes
-            model_attrs = ["model", "max_model_len", "enable_mm"]
-            for attr in model_attrs:
-                if hasattr(model_cfg, attr):
-                    self.assertTrue(hasattr(model_cfg, attr))
-
-    def test_engine_cache_config_attributes(self):
-        """Test that engine has expected cache configuration attributes"""
-        self.assertTrue(hasattr(self.engine, "cfg"))
-        self.assertIsNotNone(self.engine.cfg)
-
-        if hasattr(self.engine.cfg, "cache_config"):
-            cache_cfg = self.engine.cfg.cache_config
-            self.assertIsNotNone(cache_cfg)
-
-            # Test key cache config attributes
-            cache_attrs = ["block_size", "gpu_memory_utilization", "enable_prefix_caching"]
-            for attr in cache_attrs:
-                if hasattr(cache_cfg, attr):
-                    self.assertTrue(hasattr(cache_cfg, attr))
-
-    def test_engine_parallel_config_attributes(self):
-        """Test that engine has expected parallel configuration attributes"""
-        self.assertTrue(hasattr(self.engine, "cfg"))
-        self.assertIsNotNone(self.engine.cfg)
-
-        if hasattr(self.engine.cfg, "parallel_config"):
-            parallel_cfg = self.engine.cfg.parallel_config
-            self.assertIsNotNone(parallel_cfg)
-
-            # Test key parallel config attributes
-            parallel_attrs = ["tensor_parallel_size", "data_parallel_size", "device_ids"]
-            for attr in parallel_attrs:
-                if hasattr(parallel_cfg, attr):
-                    self.assertTrue(hasattr(parallel_cfg, attr))
-
-    def test_engine_structured_outputs_config(self):
-        """Test that engine has structured outputs configuration"""
-        self.assertTrue(hasattr(self.engine, "cfg"))
-        self.assertIsNotNone(self.engine.cfg)
-
-        if hasattr(self.engine.cfg, "structured_outputs_config"):
-            struct_cfg = self.engine.cfg.structured_outputs_config
-            self.assertIsNotNone(struct_cfg)
-
-            # Test key structured outputs config attributes
-            struct_attrs = ["guided_decoding_backend", "disable_any_whitespace", "reasoning_parser"]
-            for attr in struct_attrs:
-                if hasattr(struct_cfg, attr):
-                    self.assertTrue(hasattr(struct_cfg, attr))
-
-    def test_engine_data_processor_attributes(self):
-        """Test that engine data processor has expected attributes"""
-        if hasattr(self.engine, "data_processor"):
-            dp = self.engine.data_processor
-            self.assertIsNotNone(dp)
-
-            # Test data processor has expected attributes
-            dp_attrs = ["eos_token_id", "pad_token_id"]
-            for attr in dp_attrs:
-                if hasattr(dp, attr):
-                    self.assertTrue(hasattr(dp, attr))
-
-    def test_engine_split_connector_attributes(self):
-        """Test that engine split connector has expected attributes"""
-        if hasattr(self.engine, "split_connector"):
-            sc = self.engine.split_connector
-            self.assertIsNotNone(sc)
-
-            # Test split connector has expected attributes
-            sc_attrs = ["current_request_ids"]
-            for attr in sc_attrs:
-                if hasattr(sc, attr):
-                    self.assertTrue(hasattr(sc, attr))
-
-    def test_engine_engine_worker_queue_attributes(self):
-        """Test that engine worker queue has expected attributes"""
-        if hasattr(self.engine, "engine_worker_queue"):
-            ewq = self.engine.engine_worker_queue
-            self.assertIsNotNone(ewq)
-
-            # Test worker queue has expected methods
-            ewq_methods = ["put_tasks", "exist_tasks"]
-            for method in ewq_methods:
-                if hasattr(ewq, method):
-                    self.assertTrue(callable(getattr(ewq, method)))
-
-    def test_insert_tasks_basic_functionality(self):
-        """Test insert_tasks with real engine for basic functionality"""
-        # Create a simple request
-        request = Request(
-            request_id="test_real_1",
-            prompt="hello",
-            prompt_token_ids=[1, 2, 3],
-            prompt_token_ids_len=3,
-            messages=None,
-            history=None,
-            tools=None,
-            system=None,
-            eos_token_ids=None,
-        )
-
-        # Test basic insert functionality with real engine
-        try:
-            result = self.engine.insert_tasks([request])
-            # Should return True for successful insertion
-            self.assertIsInstance(result, bool)
-        except Exception as e:
-            # If insertion fails due to resource constraints, that's also acceptable
-            # The important thing is that the method executed without crashing
-            self.assertIsInstance(e, (EngineError, Exception))
-
-    def test_update_requests_chunk_size_real_engine(self):
-        """Test update_requests_chunk_size with real engine configuration"""
-        # Create test requests
-        request1 = Request(
-            request_id="chunk_test_1",
-            prompt="test",
-            prompt_token_ids=[1, 2, 3, 4],
-            prompt_token_ids_len=4,
-            messages=None,
-            history=None,
-            tools=None,
-            system=None,
-            eos_token_ids=None,
-        )
-
-        request2 = Request(
-            request_id="chunk_test_2",
-            prompt="test",
-            prompt_token_ids=[1, 2],
-            prompt_token_ids_len=2,
-            messages=None,
-            history=None,
-            tools=None,
-            system=None,
-            eos_token_ids=None,
-        )
-
-        # Test chunk size update with real engine
-        try:
-            self.engine.update_requests_chunk_size([request1, request2])
-            # Method should complete without error
-            self.assertTrue(True)
-        except Exception:
-            # Some configurations might not support chunking, which is fine
-            self.assertTrue(True)
-
-    def test_clear_data_real_engine(self):
-        """Test clear_data with real engine"""
-        # Test that clear_data method works with real engine
-        try:
-            result = self.engine.clear_data()
-            # Should return boolean indicating success/failure
-            self.assertIsInstance(result, bool)
-        except Exception:
-            # Clear data might fail in some configurations, which is acceptable
-            self.assertTrue(True)
-
-    def test_task_is_finished_real_engine(self):
-        """Test task_is_finished with real engine"""
-        # Test with valid indices
-        for i in range(min(10, len(self.engine.resource_manager.stop_flags))):
-            result = self.engine.task_is_finished(i)
-            self.assertIsInstance(result, (bool, np.bool_))
-
-    def test_all_tasks_finished_real_engine(self):
-        """Test all_tasks_finished with real engine"""
-        result = self.engine.all_tasks_finished()
-        self.assertIsInstance(result, (bool, np.bool_))
-
-    def test_engine_configuration_attributes(self):
-        """Test that engine has expected configuration attributes"""
-        # Test configuration attributes exist
-        self.assertTrue(hasattr(self.engine, "cfg"))
-        self.assertIsNotNone(self.engine.cfg)
-
-        # Test key configuration sections
-        config_sections = ["parallel_config", "cache_config", "model_config", "scheduler_config"]
-        for section in config_sections:
-            if hasattr(self.engine.cfg, section):
-                self.assertIsNotNone(getattr(self.engine.cfg, section))
-
-    def test_engine_resource_manager(self):
-        """Test that engine resource manager is properly initialized"""
-        if hasattr(self.engine, "resource_manager"):
-            rm = self.engine.resource_manager
-            self.assertIsNotNone(rm)
-
-            # Test resource manager has expected attributes
-            expected_attrs = ["stop_flags", "real_bsz"]
-            for attr in expected_attrs:
-                if hasattr(rm, attr):
-                    self.assertTrue(hasattr(rm, attr))
-
-    def test_engine_scheduler(self):
-        """Test that engine scheduler is properly initialized"""
-        if hasattr(self.engine, "scheduler"):
-            scheduler = self.engine.scheduler
-            self.assertIsNotNone(scheduler)
-
-            # Test scheduler has basic functionality
-            if hasattr(scheduler, "get_requests"):
-                self.assertTrue(callable(getattr(scheduler, "get_requests")))
-
-    def test_engine_token_processor(self):
-        """Test that engine token processor is properly initialized"""
-        if hasattr(self.engine, "token_processor"):
-            tp = self.engine.token_processor
-            self.assertIsNotNone(tp)
-
-            # Test token processor has expected attributes
-            expected_attrs = ["number_of_tasks", "number_of_input_tokens"]
-            for attr in expected_attrs:
-                if hasattr(tp, attr):
-                    self.assertTrue(hasattr(tp, attr))
-
-    def test_exit_sub_services_real_engine_safe_check(self):
-        """Test _exit_sub_services method exists and is callable with real engine"""
-        # This is a safe test that only checks method existence and basic attributes
-        # We don't actually call _exit_sub_services as it would shut down the real engine
-
-        # Verify the method exists
-        self.assertTrue(hasattr(self.engine, "_exit_sub_services"))
-        self.assertTrue(callable(getattr(self.engine, "_exit_sub_services")))
-
-        # Verify engine has the attributes that would be cleaned up
-        # (but don't actually clean them)
-        cleanup_related_attrs = [
-            "running",  # Engine running state
-            "cfg",  # Configuration
-        ]
-
-        for attr in cleanup_related_attrs:
-            self.assertTrue(hasattr(self.engine, attr), f"Engine missing {attr}")
-
-        # Check if engine has IPC signals (created during initialization)
-        signal_attrs = ["worker_ready_signal", "loaded_model_signal", "exist_task_signal"]
-
-        signals_present = 0
-        for attr in signal_attrs:
-            if hasattr(self.engine, attr):
-                signals_present += 1
-
-        # At least some signals should be present in a properly initialized engine
-        self.assertGreater(signals_present, 0, "Engine should have IPC signals initialized")
-
 
 if __name__ == "__main__":
     unittest.main()
@@ -1037,7 +712,6 @@ class TestCommonEngineAdditionalCoverage(unittest.TestCase):
             except Exception:
                 pass
 
-
     def test_check_worker_initialize_status_progress(self):
         """Cover 1710-1762 by simulating stdout and ready signals."""
         cfg = self._make_cfg(splitwise_role="mixed")
@@ -1181,7 +855,3 @@ class TestCommonEngineAdditionalCoverage(unittest.TestCase):
                 eng._finalizer.detach()
             except Exception:
                 pass
-
-
-if __name__ == "__main__":
-    unittest.main()
